fix: ProfileStore 따로 관리
